% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{acceptable_var_name}
\alias{acceptable_var_name}
\title{is the variable name acceptable}
\usage{
acceptable_var_name(text)
}
\arguments{
\item{text}{character string to test}
}
\value{
boolean
}
\description{
is the variable name acceptable
}
\keyword{internal}
