% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_0_side_functions.R
\name{add_coeffs}
\alias{add_coeffs}
\title{Add coefficients to the dataset
This function adds coefficients from a table to the dataframe that will be used by the solver}
\usage{
add_coeffs(
  listcoeff,
  database = t_data,
  pos.coeff.name = 1,
  pos.coeff.value = 2,
  overwrite = TRUE
)
}
\arguments{
\item{listcoeff}{data frame with at least two columns : 1 for name of coeff , another one for value of coeff}

\item{database}{target data base, Default: t_data}

\item{pos.coeff.name}{which column in list coeff has the name, Default: 1}

\item{pos.coeff.value}{which column in list coeff has the value, Default: 2}

\item{overwrite}{boolean. Whether to overwrite existing values of coefficients already in the database. Default : TRUE.}
}
\value{
returns the original dataset + the coeffs
}
\description{
Add coefficients to the dataset
This function adds coefficients from a table to the dataframe that will be used by the solver
}
