% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{check_eq_var_identification}
\alias{check_eq_var_identification}
\title{Checking if model is well identified}
\usage{
check_eq_var_identification(endo, eqns, names_of_equations)
}
\arguments{
\item{endo}{endogenous variables as declared}

\item{eqns}{equation identification matrix}

\item{names_of_equations}{vector with equation names}
}
\value{
check if identification successful
}
\description{
Checking if model is well identified
}
\keyword{internal}
