% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{check_equation_input}
\alias{check_equation_input}
\title{Basic equation checks}
\usage{
check_equation_input(equation_vector)
}
\arguments{
\item{equation_vector}{character vector to check}
}
\value{
check
}
\description{
Basic equation checks
}
\keyword{internal}
