% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{check_model_file}
\alias{check_model_file}
\title{check if external model input is ok}
\usage{
check_model_file(modelefile)
}
\arguments{
\item{modelefile}{path to file to check}
}
\value{
checks
}
\description{
check if external model input is ok
}
\keyword{internal}
