% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{check_var_vector}
\alias{check_var_vector}
\title{Checks for vectors}
\usage{
check_var_vector(variable_vector, name = "variables")
}
\arguments{
\item{variable_vector}{character vecteor of variable names to check}

\item{name}{object of the vector that is being checked}
}
\value{
check
}
\description{
Checks for vectors
}
\keyword{internal}
