% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{check_variable_conflict}
\alias{check_variable_conflict}
\title{Checking if there is a variable conflict}
\usage{
check_variable_conflict(var1, var2)
}
\arguments{
\item{var1}{First set of variables to check}

\item{var2}{Second set of variables to check}
}
\value{
checks conflict
}
\description{
Checking if there is a variable conflict
}
\keyword{internal}
