% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_9_sparse_cpp_writer.R
\name{cpp_chunked_function}
\alias{cpp_chunked_function}
\title{Split a vector of C++ statements into chunked functions}
\usage{
cpp_chunked_function(stmts, fname, sig, callargs)
}
\arguments{
\item{stmts}{character vector of statement lines}

\item{fname}{base name of the generated function}

\item{sig}{argument list of the generated function}

\item{callargs}{arguments forwarded to each chunk}
}
\value{
character vector of C++ lines defining \code{fname} and its chunks
}
\description{
Split a vector of C++ statements into chunked functions
}
\keyword{internal}
