% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_0_single_equations.R
\name{create_equation}
\alias{create_equation}
\title{Creating a thoR.equation}
\usage{
create_equation(
  equation_name = "my_equation",
  formula,
  coefflist = NULL,
  endogenous = NULL,
  env = globalenv(),
  default_maxlag_error = 5
)
}
\arguments{
\item{equation_name}{character string to name the equations}

\item{formula}{character string of full formula}

\item{coefflist}{coefficient variables for information. Create equation will try to create a RHS and LHS that is compatible with econometric estimations}

\item{endogenous}{the endogenous variable of the equation (only one)}

\item{env}{environment where the the equation should be loaded}

\item{default_maxlag_error}{number of lag to default to if the maximum lag of the equation cannot be determined}
}
\value{
thoR.equation object
}
\description{
Creating a thoR.equation
}
