% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{create_equations_list}
\alias{create_equations_list}
\title{Create the equations_list table from an eqlist}
\usage{
create_equations_list(eq_list)
}
\arguments{
\item{eq_list}{vector of equations as raw input}
}
\value{
data.frame equations lists
}
\description{
Create the equations_list table from an eqlist
}
\keyword{internal}
