% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_7_2_create_rcpp_source_nonsuperlu.R
\name{create_model_rcpp_source_nonsuperlu}
\alias{create_model_rcpp_source_nonsuperlu}
\title{Generate the rcpp file}
\usage{
create_model_rcpp_source_nonsuperlu(
  model_name,
  p_h_e_bool,
  p_h_e_jac,
  all_model_vars,
  rcpp_path
)
}
\arguments{
\item{model_name}{name of the model}

\item{p_h_e_bool}{vector of 3 booleans for prologue, heart, epilogue}

\item{p_h_e_jac}{list of 3 jacobians for prologue, heart, epilogue}

\item{all_model_vars}{all variables that are in the model, in alphabetical order}

\item{rcpp_path}{path where to store the rcpp file}
}
\value{
rcpp source file
}
\description{
Generate the rcpp file
}
\author{
Charlotte Nudelmann (DG Tresor)
}
\keyword{internal}
