% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_data_checks.R
\name{data_model_checks}
\alias{data_model_checks}
\title{Checking if all needed variables is in the data}
\usage{
data_model_checks(thor_model, database, quiet = TRUE)
}
\arguments{
\item{thor_model}{a thoR.model object or thoR.equation}

\item{database}{the database to check}

\item{quiet}{Display message if successful. Default : TRUE}
}
\value{
boolean : TRUE if all ok, FALSE if there was a problem
}
\description{
Checking if all needed variables is in the data
}
