% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_3_decomposition_algo.R
\name{decomposing_model}
\alias{decomposing_model}
\title{Algorithms to determine which equations and variables are part of what}
\usage{
decomposing_model(endogenous_variables, eq_var_matrix, decomposition = FALSE)
}
\arguments{
\item{endogenous_variables}{vector containing all the declred endogenous variables}

\item{eq_var_matrix}{matrix of contemporaneous endogenous variables per equations}

\item{decomposition}{boolean, TRUE if go ahead with decomposition}
}
\value{
list containing all parts infos
}
\description{
Algorithms to determine which equations and variables are part of what
}
\keyword{internal}
