% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_4_dates.R
\name{detect_frequency}
\alias{detect_frequency}
\title{Detect frequency}
\usage{
detect_frequency(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates (class "Date").}
}
\value{
one of (undetermined,daily, weekly, monthly, quarterly, biannual or annual)
}
\description{
Detect frequency
}
\keyword{internal}
