% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_export.R
\docType{data}
\name{dictionnaire_variables_opale}
\alias{dictionnaire_variables_opale}
\title{list of functions supported by the parser}
\format{
character
}
\source{
Dictionnaire des variables d'Opale
}
\usage{
data(dictionnaire_variables_opale)

##Dataexport
}
\description{
list of functions supported by the parser
}
\author{
Anissa
}
\keyword{data}
\keyword{opale}
