% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_0_visual_outputs.R
\name{dyn_contribs}
\alias{dyn_contribs}
\title{Dynamic contributions for ECM equations}
\usage{
dyn_contribs(
  thor_equation,
  database,
  start_sim,
  end_sim,
  index_time = "date",
  residual_var,
  regroup_these = NULL,
  name_group = "group",
  print_verif = TRUE
)
}
\arguments{
\item{thor_equation}{thoR.equation object, must be an error correction model type of equation, otherwise results wont be accurate.}

\item{database}{data.frame with data for all variables in the equation.}

\item{start_sim}{Start period to start the simulations.  Same class as used in index_time. Must be in index_time.}

\item{end_sim}{Last period to end the simulations.  Same class as used in index_time. Must be in index_time.}

\item{index_time}{Name of the column in the database which is meant to be use as the time vector.}

\item{residual_var}{character string, length 1. Variable of the equation that serves as the residual var.}

\item{regroup_these}{character vector of variables which contributions will be summed. Default: NULL}

\item{name_group}{character string to name the grouped variables contribution. Default: "group"}

\item{print_verif}{boolean. TRUE if verification of contributions and residuals should be printed. Default: TRUE}
}
\value{
data.frame containing dynamic contributions as well as simulation results
}
\description{
Dynamic contributions for ECM equations
}
