% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_5b_sparse_jacobian.R
\name{empty_sparse_jacobian}
\alias{empty_sparse_jacobian}
\title{An empty sparse jacobian, for blocks that are not used}
\usage{
empty_sparse_jacobian()
}
\description{
An empty sparse jacobian, for blocks that are not used
}
\keyword{internal}
