% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_equations_functions.R
\name{equation_from_model}
\alias{equation_from_model}
\title{Pull an equation from a model}
\usage{
equation_from_model(
  equation_name_or_id,
  model,
  endogenous,
  new_name = NULL,
  default_maxlag_error = 5,
  environment = globalenv()
)
}
\arguments{
\item{equation_name_or_id}{character string of length 1 with the equation name or id, must match the name or id used in the model.}

\item{model}{thoR.model object where to pull the equation from}

\item{endogenous}{character string of length 1 that is a variable of the equation and will be used as the equation's endogenous variable.}

\item{new_name}{character string to designate the new name of the model. if NULL, the name will be the same as the equation_name_or_id. Default : NULL.}

\item{default_maxlag_error}{if maxlag fails it will default to this value. Default : 5.}

\item{environment}{environment where the equation created must be loaded. Default : globalenv()}
}
\value{
a thoR.equation from a thoR.model
}
\description{
This function returns a thoR.equation object created from an equation in a thoR.model.
}
