% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_0_model_analysis.R
\name{equation_info_model}
\alias{equation_info_model}
\title{Information about equations in the model}
\usage{
equation_info_model(equations_name_or_id, model)
}
\arguments{
\item{equations_name_or_id}{equations names or id in a character vector.}

\item{model}{thoR.model with the equations}
}
\value{
Prints information about the equations
}
\description{
Information about equations in the model
}
