% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_files_and_data_management.R
\name{export_model}
\alias{export_model}
\title{Export the model as a .txt file}
\usage{
export_model(model, filename = "model.txt")
}
\arguments{
\item{model}{thoR.model obect}

\item{filename}{complete path and name of file to save}
}
\value{
Saved file
}
\description{
Export the model as a .txt file
}
