% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_7_rcpp_source_builder.R
\name{f_x_Rcpp}
\alias{f_x_Rcpp}
\title{Transforming the equations R source code into Rcpparmadillo code}
\usage{
f_x_Rcpp(path_f_x, var_order, cpp_file = NA)
}
\arguments{
\item{path_f_x}{where and which equation source file}

\item{var_order}{all model variables in alphabetical order}

\item{cpp_file}{name of cpp file where to write}
}
\value{
cpp script
}
\description{
Transforming the equations R source code into Rcpparmadillo code
}
\author{
Charlotte Nudelmann (DG Tresor)
}
\keyword{internal}
