% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_equations_functions.R
\name{find_maxlag}
\alias{find_maxlag}
\title{determining the maximum lag an equation with econometric format (newdiff instead of delta)
the formula must have been formatted to be evaluated by thor first}
\usage{
find_maxlag(new_formula_vector, default_error = NA)
}
\arguments{
\item{new_formula_vector}{vector containing equations with the format that can be evaluated}

\item{default_error}{default value in case of error in algorithm. Default : NA}
}
\value{
maxlag value or NA if couldn't determine it.
}
\description{
determining the maximum lag an equation with econometric format (newdiff instead of delta)
the formula must have been formatted to be evaluated by thor first
}
\keyword{internal}
