% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{formatting_formulas}
\alias{formatting_formulas}
\title{Formatting formula
rearranges logs and lags to fit  derivators and such.}
\usage{
formatting_formulas(formula_list)
}
\arguments{
\item{formula_list}{vector of formula to be modified}
}
\value{
character vector of reformatted formulas easy for computation of jacobians
}
\description{
Formatting formula
rearranges logs and lags to fit  derivators and such.
}
\keyword{internal}
