% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_4_latex_formula.R
\name{formula_latex}
\alias{formula_latex}
\title{Transform a formula into LaTex code for display}
\usage{
formula_latex(equation, coeff_list = NULL)
}
\arguments{
\item{equation}{a formula as a character string or a thoR.equation}

\item{coeff_list}{a vector of coeeficients that could be in the equation. If equation is a thoR.equation object, this argument is skipped. If NULL coefficients that are still presnet in the formula will not be formatted.}
}
\value{
a character string written into LaTex to use with Rmarkdown or LaTex
}
\description{
Transform a formula into LaTex code for display
}
\examples{
my_eq <- "delta(1,log(var_1)) = cst + cf_1 *delta(1,log(var_2)) - 0.5 * lag(var_1,1)"
cat(formula_latex(my_eq,c("cst","cf1")))
}
