% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_equations_functions.R
\name{formula_with_coeffs}
\alias{formula_with_coeffs}
\title{Shows the formula with the coefficients' value in numeric instead of the coefficient variable names.}
\usage{
formula_with_coeffs(
  formula,
  coefflist = NULL,
  database,
  round_digits = 4,
  quiet = FALSE
)
}
\arguments{
\item{formula}{a formula as a character string or a thor equation}

\item{coefflist}{vector with variables that could be coefficients. If the formula input is a thoR.equation model, this argument is ignored, and the coefficients are taken from the object parameters}

\item{database}{database that contains the values of the formula coefficient. Name of coefficient must be as column names and values the same across all observations of the base. The first row will be used. Named vector should also work.}

\item{round_digits}{Numeric. numbers of digits to round the value of the coefficient. Default : 4}

\item{quiet}{boolean. FALSE doesn't display the result. Default : TRUE}
}
\value{
character string
}
\description{
Shows the formula with the coefficients' value in numeric instead of the coefficient variable names.
}
