% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{generate_equations_index}
\alias{generate_equations_index}
\title{Generating equation index id}
\usage{
generate_equations_index(equation_list)
}
\arguments{
\item{equation_list}{a vector}
}
\value{
character vector with eq_###
}
\description{
Generating equation index id
}
\keyword{internal}
