% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_0_fonctions_estimation.R
\name{get_coeff_lt_value}
\alias{get_coeff_lt_value}
\title{Estimating long term part of ECM}
\usage{
get_coeff_lt_value(
  var_lt,
  var_log_all,
  endogenous_name,
  coeff_lt,
  sub_database,
  estim_start,
  estim_end,
  index_time = "date"
)
}
\arguments{
\item{var_lt}{plop}

\item{var_log_all}{plop}

\item{endogenous_name}{plop}

\item{coeff_lt}{plop}

\item{sub_database}{plop}

\item{estim_start}{plop}

\item{estim_end}{plop}

\item{index_time}{plop}
}
\value{
data.frame with long-term coefficients values
}
\description{
Estimating long term part of ECM
}
\keyword{internal}
