% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_4_function_command_writers.R
\name{get_functions_commands}
\alias{get_functions_commands}
\title{Generate the lines of the function for evaluating equations
stored as a vector for easy manipulation}
\usage{
get_functions_commands(
  list_of_equations,
  database_name = "t_data",
  output_vector_name = "f_x_n",
  currentstep = "prologue_equations"
)
}
\arguments{
\item{list_of_equations}{vector of equations}

\item{database_name}{database of the solver (t_data)}

\item{output_vector_name}{vector name as used in solver f_x_n}

\item{currentstep}{will be used in the function name}
}
\value{
character vector with the lines to build the function
}
\description{
Generate the lines of the function for evaluating equations
stored as a vector for easy manipulation
}
\keyword{internal}
