% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_4_function_command_writers.R
\name{get_jacobian_commands}
\alias{get_jacobian_commands}
\title{Generate the command lines of function to evaluate the jacobian}
\usage{
get_jacobian_commands(
  jacobian,
  database_name,
  output_matrix_name = "output",
  info.matrix = "jacobian",
  currentstep = "prologue_jacobian"
)
}
\arguments{
\item{jacobian}{jacobian symbolic matrix}

\item{database_name}{database of the solver (t_data)}

\item{output_matrix_name}{output matrix name as used in solver}

\item{info.matrix}{where to source the matrix}

\item{currentstep}{will be used in the function name}
}
\value{
character vector with the lines to build the function
}
\description{
Generate the command lines of function to evaluate the jacobian
}
\keyword{internal}
