% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_0_fonctions_estimation.R
\name{get_var_log_all}
\alias{get_var_log_all}
\title{Variables with or without logs}
\usage{
get_var_log_all(formula_txt, variables_list)
}
\arguments{
\item{formula_txt}{plop}

\item{variables_list}{plop}
}
\value{
dataframe with variables that need to be transformed  by log
}
\description{
Variables with or without logs
}
\keyword{internal}
