% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_0_fonctions_estimation.R
\name{get_var_lt}
\alias{get_var_lt}
\title{get_var_lt}
\usage{
get_var_lt(formula_txt, coeff_lt)
}
\arguments{
\item{formula_txt}{character string of formula}

\item{coeff_lt}{character vector of coefficients}
}
\description{
get_var_lt
}
\keyword{internal}
