% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{get_variables_from_string}
\alias{get_variables_from_string}
\title{Determine all variables present in a formula}
\usage{
get_variables_from_string(string)
}
\arguments{
\item{string}{character string of the formula}
}
\value{
character vector with all variables detected in the formula.
}
\description{
Determine all variables present in a formula
}
