% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_2_graphs.R
\name{graph_contrib}
\alias{graph_contrib}
\title{Draw contributions graph}
\usage{
graph_contrib(
  contrib_data,
  start_plot = NULL,
  end_plot = NULL,
  index_time = NULL,
  title = "Contributions dynamiques",
  name_endogenous = "main variable",
  colour_line = "black",
  colours_bar = NULL,
  bar_labels = NULL,
  legend_n_row = 2
)
}
\arguments{
\item{contrib_data}{database generated by dyn_contrib or the annualised version from yearly_contrib}

\item{start_plot}{start date, must be in index time. The subdata must not contain NAs.}

\item{end_plot}{end date, must be in index time. The subdata must not contain NAs.}

\item{index_time}{Name of the column which contains the time info.}

\item{title}{Title of the graph}

\item{name_endogenous}{Name of the endogenous variables represented by dlog_obs(.y) for display only.Default : "main variable"}

\item{colour_line}{One colour to customize the line colour. Default: "black".}

\item{colours_bar}{Vector of colours to customize the contrib bars, must be at least the number of contrib variables. The last colour is always the residual colour. If not enough colours are provided, default colours will be used instead. Default: NULL.}

\item{bar_labels}{Character vector to rename the contribution variables in the legend. The order must match the order of variables in the contrib database, with the exception of the residual which must be specified last. If the wrong number of labels is supplied, labels revert to the default ones.  Default: NULL.}

\item{legend_n_row}{Numeric. Number of rows the legend should be spread over. Enables customizing of the display so that the legend fits different sizes of the panel, especially when specifying different longer labels. Default: 2}
}
\value{
a ggplot graph
}
\description{
Draw contributions graph
}
