% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_2_graphs.R
\name{graph_sim_obs}
\alias{graph_sim_obs}
\title{Plot simulated vs observed}
\usage{
graph_sim_obs(
  sim_obs_data,
  start_plot = NULL,
  end_plot = NULL,
  index_time = NULL,
  title = "",
  type = c("g", "dlog", "d", "lvl"),
  colours = c("navyblue", "darkorange3")
)
}
\arguments{
\item{sim_obs_data}{data.frame generated by simulate_equation() or dyn_contrib()}

\item{start_plot}{which date should the plot start. Must be in index time. The subdata must not contain NAs.}

\item{end_plot}{end date, must be in index time. The subdata must not contain NAs.}

\item{index_time}{Name of the column which contains the time info. By default is NULL and takes the first column of sim_obs_data}

\item{title}{Title of the graph}

\item{type}{one of 'g' (growth rate), 'dlog' (diff(log)), 'd' (diff) or 'lvl'  (level)}

\item{colours}{vector of two colours. Default : c("navyblue","darkorange3")}
}
\value{
a ggplot graph
}
\description{
Plot simulated vs observed
}
