% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{is_in_formulas}
\alias{is_in_formulas}
\title{which variables in the formula}
\usage{
is_in_formulas(variables_to_test, formulas, print_type = "")
}
\arguments{
\item{variables_to_test}{character vector of variables to test}

\item{formulas}{character vector of formulas}

\item{print_type}{display the variable types that are being tested}
}
\value{
vector of variables that have been found in the formula
}
\description{
which variables in the formula
}
\keyword{internal}
