% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_7_rcpp_source_builder.R
\name{Jacobian_Rcpp}
\alias{Jacobian_Rcpp}
\title{Transforming the jacobian R source code into Rcpparmadillo code}
\usage{
Jacobian_Rcpp(
  path_jacobian_R,
  var_order = var_list_sorted,
  index_jacobienne,
  cpp_file = NA
)
}
\arguments{
\item{path_jacobian_R}{where and which jacobian}

\item{var_order}{all variables in alphabetical order}

\item{index_jacobienne}{symbolic jacobian}

\item{cpp_file}{name and location of the file.}
}
\value{
cpp script
}
\description{
Transforming the jacobian R source code into Rcpparmadillo code
}
\author{
Charlotte Nudelmann (DG Tresor)
}
\keyword{internal}
