% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{list_2DF_withNA}
\alias{list_2DF_withNA}
\title{Getting the old style list2DF}
\usage{
list_2DF_withNA(xlist)
}
\arguments{
\item{xlist}{a list}
}
\value{
dataframe with NAs when the length of the elements differ
}
\description{
Getting the old style list2DF
}
\keyword{internal}
