% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_0_model_modifications.R
\name{model_equations_add}
\alias{model_equations_add}
\title{Adds thor.equations to an existing model and builds the model.}
\usage{
model_equations_add(
  base_model,
  new_model_name,
  thor_equations_add = NULL,
  algo = TRUE,
  rcpp = FALSE,
  env = globalenv(),
  rcpp_path = getwd(),
  use.superlu = FALSE
)
}
\arguments{
\item{base_model}{: the reference model to be used (obviously a thoR.model)}

\item{new_model_name}{: character string, must be different from base model name}

\item{thor_equations_add}{list of thoR equations to add.}

\item{algo}{boolean. TRUE if using decomposition of model (recommended for large model). Default : TRUE}

\item{rcpp}{boolean. TRUE if using Rcpp for solver (recommended for large model). Default : FALSE}

\item{env}{Environment where to store the model object. Default : globalenv()}

\item{rcpp_path}{path to directory where to store the rcpp source files for the model. Default : working directory}

\item{use.superlu}{boolean. If SUPERLU library is installed, select TRUE to compile a rcpp model with superlu}
}
\value{
a thor.model in the selected environment, also returned invisibly
}
\description{
Adds thor.equations to an existing model and builds the model.
}
