% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_files_and_data_management.R
\name{model_source_example}
\alias{model_source_example}
\title{load a sample model file}
\usage{
model_source_example(path_only = FALSE)
}
\arguments{
\item{path_only}{TRUE to disable browsing and only show the path instead. Default: FALSE}
}
\value{
Opens a sample model file or the path only if path_only = TRUE.
}
\description{
load a sample model file
}
\examples{
\dontrun{model_source_example()}
model_source_example(TRUE)
}
