% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_NA_check.R
\name{na_report_variables_times}
\alias{na_report_variables_times}
\title{Report which variables have NAs for given periods}
\usage{
na_report_variables_times(database, times, variables, index_time = "date")
}
\arguments{
\item{database}{a dataframe containing all necessary variables}

\item{times}{periods in index_times to test}

\item{variables}{variables to check}

\item{index_time}{name of column in the database which is meant to be use as the time vector}
}
\value{
vector of variable names with NAs over the time period
}
\description{
Report which variables have NAs for given periods
}
