% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_0_side_functions.R
\name{newdiff}
\alias{newdiff}
\alias{delta}
\title{compute n order differences while keeping the vector same length as original one}
\usage{
newdiff(n = 1, x)

delta(n = 1, x)
}
\arguments{
\item{n}{number of periods to diff}

\item{x}{numeric vector of variable to diff}
}
\value{
numeric vector of diff
}
\description{
compute n order differences while keeping the vector same length as original one
}
