% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_check_model_files.R
\name{parser_lag_delta_check}
\alias{parser_lag_delta_check}
\title{check if the lags and delta can be safely parser}
\usage{
parser_lag_delta_check(formula_list)
}
\arguments{
\item{formula_list}{vector of formulas}
}
\value{
boolean
}
\description{
check if the lags and delta can be safely parser
}
\keyword{internal}
