% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_0_fonctions_estimation.R
\name{quick_estim}
\alias{quick_estim}
\title{Quickly estimate an OLS or ECM equation from a model}
\usage{
quick_estim(
  thor_equation = NULL,
  eq_name = NULL,
  thor_model = NULL,
  endogenous_name = NULL,
  database,
  index_time = "date",
  estim_start = NULL,
  estim_end = NULL,
  coeff_lt = NULL,
  const = TRUE
)
}
\arguments{
\item{thor_equation}{a thoR.equation object. Default : NULL}

\item{eq_name}{If no thor_equation is specified, name of the equation in the thor_model. Default : NULL}

\item{thor_model}{If no thor_equation is specified, the thoR.model where to get the equation eq_name. Default : NULL}

\item{endogenous_name}{If no thor_equation is specified, endogenous variable of the equation, must be a variable of the equation. Default : NULL}

\item{database}{data.frame with the necessary data to estimate the equation. Default : NULL}

\item{index_time}{name of column in the database which is meant to be use as the time vector. Default : "date"}

\item{estim_start}{First period of estimation, must be in index_time. Default : NULL}

\item{estim_end}{Last period of estimation, must be in index_time.. Default : NULL}

\item{coeff_lt}{list of long term coefficients. If NULL then the equation will only be estimated as OLS (one step). Default : NULL}

\item{const}{FALSE if the constant is not to be estimated. Default : TRUE}
}
\value{
database with the coefficients with all coefficients estimated
}
\description{
Quickly estimate an OLS or ECM equation from a model
}
\author{
Niamh Dunne
}
