% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_0_fonctions_estimation.R
\name{quick_estim_all}
\alias{quick_estim_all}
\title{Quickly estimate multiple equations (OLS or ECM) from a model}
\usage{
quick_estim_all(info_equations, thor_model, database, index_time)
}
\arguments{
\item{info_equations}{List (level 2) with the information about the equations. Each element (level 1) of the list must be a list named after an equation name of the model, that contains at least the following elements which correspond to the arguments of the \code{quick_estim()} function: endogenous_name ,coeff_lt, estim_start, estim_end and const.}

\item{thor_model}{thoR.model that contains the}

\item{database}{data.frame with the necessary data to estimate all the equations in the list. Default : NULL}

\item{index_time}{name of column in the database which is meant to be use as the time vector. Default : "date"}
}
\value{
database with the coefficients with all coefficients estimated
}
\description{
Quickly estimate multiple equations (OLS or ECM) from a model
}
\examples{
\dontrun{See the UK_exemple vignette for a comprehensive exemple.}
}
\author{
Niamh Dunne
}
