% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_3_quick_solve.R
\name{quick_solve}
\alias{quick_solve}
\title{Quickly solve a static equation}
\usage{
quick_solve(formula = "5*x=210", endogenous = "x", init = 1, quiet = FALSE)
}
\arguments{
\item{formula}{equation formula, must not have any symbolic variable other than the endogenous variable. Default: "5*x=210"}

\item{endogenous}{Variable to solve as it appears in the equation. Default: "x"}

\item{init}{numeric initial value for the solver to start the solving algorithm. Default : 1}

\item{quiet}{TRUE to not display the result in the console. Default : FALSE}
}
\value{
a numeric value, a possible solution of the equation
}
\description{
Quickly solve a static equation
}
\examples{
quick_solve("log(x)+x=10","x")
quick_solve("3*x+4=10","x")
}
