% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_4_dates.R
\name{reformat_date}
\alias{reformat_date}
\title{Reformat a date vector based on its frequency}
\usage{
reformat_date(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates (class "Date").}
}
\value{
a character vector made to look nicer for graphs
}
\description{
Reformat a date vector based on its frequency
}
\examples{
date<-seq.Date(from=as.Date("2019-02-04"),by = "quarter",length.out = 10)
print(reformat_date(date))
}
