% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_equations_functions.R
\name{reorder_coefficients}
\alias{reorder_coefficients}
\title{Prepares the RHS of the equation and reorder coefficient list for econometric estimation}
\usage{
reorder_coefficients(LHSformula, RHSformula, coeffs_variables, eq_vars)
}
\arguments{
\item{LHSformula}{left hand side of formula}

\item{RHSformula}{right hand side of formula}

\item{coeffs_variables}{list of coefficients}

\item{eq_vars}{equation variables}
}
\value{
list with modified RHS and coefficients in the right order
}
\description{
Prepares the RHS of the equation and reorder coefficient list for econometric estimation
}
\keyword{internal}
