% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_files_and_data_management.R
\name{save_model}
\alias{save_model}
\title{Save a thoR model
The model is saved as a .rds file.}
\usage{
save_model(model, folder_path = getwd())
}
\arguments{
\item{model}{thor.model object to save.}

\item{folder_path}{directory where to store the .rds. Default : working directory}
}
\value{
rds file stored in location
}
\description{
Save a thoR model
The model is saved as a .rds file.
}
\examples{
\dontrun{
save_model(forecast_model)
}
}
