% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_0_visual_outputs.R
\name{simulate_equation}
\alias{simulate_equation}
\title{Simulate an equation}
\usage{
simulate_equation(
  thor_equation,
  database,
  start_sim,
  end_sim,
  index_time = "date",
  residual_var
)
}
\arguments{
\item{thor_equation}{thoR.equation object}

\item{database}{data.frame with data for all variables in the equation.}

\item{start_sim}{Start period to start the simulations.  Same class as used in index_time. Must be in index_time.}

\item{end_sim}{Last period to end the simulations.  Same class as used in index_time. Must be in index_time.}

\item{index_time}{Name of the column in the database which is meant to be use as the time vector.}

\item{residual_var}{character string, length 1. Variable of the equation that serves as the residual var.}
}
\value{
data.frame with the equation variable data. The "observed" variable represent the endogenous variable before simulation. The "simulated" variable is after estimation with the residual variable set 0. The residual is also estimated, as well as its contribution.
}
\description{
Simulate an equation
}
