% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_5_symbolic_jacobian.R
\name{symbolic_jacobian}
\alias{symbolic_jacobian}
\title{Computing the symbolic jacobian}
\usage{
symbolic_jacobian(
  equations_list_df,
  eqns_vars_list,
  endo_vec,
  equations_subset,
  id_col = "id",
  formula_col = "new_formula"
)
}
\arguments{
\item{equations_list_df}{dataframe with the equation infos and formulas.}

\item{eqns_vars_list}{equations variable matrix as a list.}

\item{endo_vec}{vector of the endogenous variables by which to compute the partial derivative.}

\item{equations_subset}{vector with the ids of the equations.}

\item{id_col}{column in equations_list_df that contains the id. Default : "id"}

\item{formula_col}{column in equations_list_df that contains the formula. Default : "new_formula"}
}
\value{
character, named matrix of jacobians
}
\description{
Computing the symbolic jacobian
}
\keyword{internal}
