% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_create_model_subfunctions.R
\name{table_contemporaneous_endos}
\alias{table_contemporaneous_endos}
\title{Creating the table of contemporenous endos}
\usage{
table_contemporaneous_endos(
  formula_list,
  endogenous,
  exogenous,
  coefflist,
  equations_index,
  functionsthor = thor_functions_supported
)
}
\arguments{
\item{formula_list}{vector containing all formulas}

\item{endogenous}{vector of endogenous variables}

\item{exogenous}{vector of exogenous variableq}

\item{coefflist}{vector of coefficient}

\item{equations_index}{vector of equation index}

\item{functionsthor}{vector of supported functions : defaults to thor_functions_supported}
}
\value{
data.frame with all contemporaneous endos per equation
}
\description{
Creating the table of contemporenous endos
}
\keyword{internal}
