% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_thor_classes.R
\docType{class}
\name{thoR.model-class}
\alias{thoR.model-class}
\alias{thoR.model}
\title{thoR models with multiple equations
Models are created using the create_model function}
\description{
thoR models with multiple equations
Models are created using the create_model function
}
