% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_1_thor_single_solver.R
\name{thor_equation_solver}
\alias{thor_equation_solver}
\title{thor solver}
\usage{
thor_equation_solver(
  equation,
  first_period,
  last_period,
  database = t_data,
  index_time = "date",
  convergence_criteria = 1e-13
)
}
\arguments{
\item{equation}{A thoR.equation}

\item{first_period}{First period to solve. Same class as used in index_time. Must be in index_time.}

\item{last_period}{Last period to solve. Same class as used in index_time. Must be in index_time.}

\item{database}{data.frame database to be used. default is t_data}

\item{index_time}{name of column in the database which is meant to be use as the time vector}

\item{convergence_criteria}{initial convergence criteria of the model. Default : 10e^-16}
}
\value{
a data.frame with only the variables in the equation. The endogenous solved appears as endogenous.simul
}
\description{
thor solver
}
