% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_export.R
\docType{data}
\name{thor_functions_supported}
\alias{thor_functions_supported}
\title{Functions Supported}
\format{
character
}
\source{
Based on Deriv package functions
}
\usage{
data(thor_functions_supported)

##Dataexport
}
\description{
list of functions supported by the parser
}
\author{
Anissa
}
\keyword{data}
