% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_0_solver.R
\name{thor_solver}
\alias{thor_solver}
\title{thor solver}
\usage{
thor_solver(
  model,
  first_period,
  last_period,
  main_variable = NULL,
  main_variable_fx = function(x) x,
  database = t_data,
  index_time = "date",
  convergence_criteria = 1e-09,
  dynamic_convergence = 10,
  rcpp = FALSE,
  adv_diag = FALSE
)
}
\arguments{
\item{model}{A thoR.model}

\item{first_period}{First period to solve. Same class as used in index_time. Must be in index_time.}

\item{last_period}{Last period to solve. Same class as used in index_time. Must be in index_time.}

\item{main_variable}{Main variable of interest for quick preview of the info.Must be a variable of the dataframe.}

\item{main_variable_fx}{function to aply to the main variable displayed. Default : function(x)x}

\item{database}{data.frame database to be used. default is t_data}

\item{index_time}{name of column in the database which is meant to be use as the time vector}

\item{convergence_criteria}{initial convergence criteria of the model. Default : 10e^-09}

\item{dynamic_convergence}{numeric/integer. number of iterations before the solver increases the convergence criteria by a factor of 10. Default : 10}

\item{rcpp}{boolean. If the rcpp solver is to be used. Default : FALSE}

\item{adv_diag}{: boolean : gives advanced informations aout the computations. Default : FALSE}
}
\value{
a dataframe that mirrors the database input, but with endogenous variables solved.
}
\description{
thor solver
}
