% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_NA_check.R
\name{time_solver_test_run}
\alias{time_solver_test_run}
\title{Tests if the solver has enough data to run
useful to check the need of lag datas}
\usage{
time_solver_test_run(model, database, index_time = "date", times)
}
\arguments{
\item{model}{a thoR.model object or thoR.equation}

\item{database}{a dataframe containing all necessary variables}

\item{index_time}{name of column in the database which is meant to be use as the time vector}

\item{times}{periods to test.}
}
\value{
boolean, TRUE if successful
}
\description{
Tests if the solver has enough data to run
useful to check the need of lag datas
}
