% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_0_model_analysis.R
\name{var_info_model}
\alias{var_info_model}
\title{Information about variables in the model.}
\usage{
var_info_model(variables, model)
}
\arguments{
\item{variables}{character vector of variables to look up.}

\item{model}{thoR.model of interest}
}
\value{
Prints information about the variables : type and the equations where they appear.
}
\description{
Information about variables in the model.
}
