% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_0_visual_outputs.R
\name{yearly_contrib}
\alias{yearly_contrib}
\title{Quartely to yearly dynamic contributions}
\usage{
yearly_contrib(contrib_quarterly_database, index_year, quarter_start = 1)
}
\arguments{
\item{contrib_quarterly_database}{quarterly dynamic contributions data generated by dyn_contrib()}

\item{index_year}{year vector that matches the observations in the data}

\item{quarter_start}{numeric. Which is the first quarter of the database (1,2,3 or 4). Default: 1.}
}
\value{
annual database with yearly dynamic contribution
}
\description{
Transforms  quarterly dynamic contribution data into yearly data, ready for graphics
}
