% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_create_model.R
\name{create_model}
\alias{create_model}
\title{Create a thoR.model object}
\usage{
create_model(
  model_name = "model",
  model_source = NULL,
  algo = TRUE,
  rcpp = FALSE,
  rcpp_path = getwd(),
  endogenous = NULL,
  exogenous = NULL,
  coefficients = NULL,
  equations = NULL,
  env = globalenv(),
  no_var_map = FALSE
)
}
\arguments{
\item{model_name}{character. Name of the model object. Default: 'prev'}

\item{model_source}{path. path to the .txt file that contains the model info. Default: 'NULL'}

\item{algo}{boolean. TRUE if using decomposition of model (recommended for large model). Default : TRUE}

\item{rcpp}{boolean. TRUE if using Rcpp for solver (recommended for large model). Default: FALSE}

\item{rcpp_path}{path to directory where to store the rcpp source files for the model. Default: working directory}

\item{endogenous}{character vector of endogenous variables (only read if no input is provided in model_source). Default: NULL}

\item{exogenous}{character vector of exogenous variables (only read if no input is provided in model_source, optionnal). Default: NULL}

\item{coefficients}{character vector of coefficients (only read if no input is provided in model_source, optionnal). Default: NULL}

\item{equations}{character vector containing equations (only read if no input is provided in model_source, optionnal). Default: NULL}

\item{env}{Environment where to store the model object. Default: globalenv()}

\item{no_var_map}{boolean. If TRUE, no vat_map we will be built, instead an empty list will be created. Default : FALSE}
}
\value{
A thoR.model created in the environment.
}
\description{
This function creates a thoR.model object by reading a model in text file or specified from a manual input in the function arguments.
}
\examples{
\dontrun{
 create_model("My_Model",model_source_example(TRUE))
 create_model("Opale",system.file("models", "opale.txt", package = "tresthor"),rcpp=TRUE)
 }

}
