% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0_files_and_data_management.R
\name{load_model}
\alias{load_model}
\title{Load a saved model}
\usage{
load_model(
  model = NULL,
  folder_path = getwd(),
  file = NULL,
  env = globalenv(),
  source_rcpp = TRUE
)
}
\arguments{
\item{model}{string containing model name (without any extension)}

\item{folder_path}{directory where to find the model}

\item{file}{Alternatively provide the complete file path, with the extension. this argument will take over model + folder path argument.}

\item{env}{environment where to load the model. Default : globalenv()}

\item{source_rcpp}{Boolean. TRUE if sourcing of the rcpp file is needed}
}
\value{
loaded model in the environment, ready to be solved.
}
\description{
Load a saved model
}
